+++ js/iosrtc.js
@@ -23,8 +23,9 @@ var // Dictionary of MediaStreamRenderers.
 	RTCIceCandidate = require('./RTCIceCandidate'),
 	MediaDevices = require('./MediaDevices'),
 	MediaStream = require('./MediaStream'),
-	MediaStreamTrack = require('./MediaStreamTrack'),
-	videoElementsHandler = require('./videoElementsHandler');
+	{ MediaStreamTrack } = require('./MediaStreamTrack'),
+	videoElementsHandler = require('./videoElementsHandler'),
+	{ RTCRtpTransceiver }	   = require('./RTCRtpTransceiver');
 
 /**
  * Expose the iosrtc object.
@@ -217,6 +218,7 @@ function registerGlobals(doNotRestoreCallbacksSupport) {
 	window.MediaStream = MediaStream;
 	window.webkitMediaStream = MediaStream;
 	window.MediaStreamTrack = MediaStreamTrack;
+	window.RTCRtpTransceiver = RTCRtpTransceiver;
 
 	// Apply CanvasRenderingContext2D.drawImage monkey patch
 	var drawImage = CanvasRenderingContext2D.prototype.drawImage;