+++ lib/WebRTC.xcframework/ios-i386_x86_64-simulator/WebRTC.framework/Headers/RTCSessionDescription.h
@@ -20,12 +20,13 @@ typedef NS_ENUM(NSInteger, RTCSdpType) {
   RTCSdpTypeOffer,
   RTCSdpTypePrAnswer,
   RTCSdpTypeAnswer,
+  RTCSdpTypeRollback,
 };
 
 NS_ASSUME_NONNULL_BEGIN
 
 RTC_OBJC_EXPORT
-@interface RTCSessionDescription : NSObject
+@interface RTC_OBJC_TYPE (RTCSessionDescription) : NSObject
 
 /** The type of session description. */
 @property(nonatomic, readonly) RTCSdpType type;