+++ lib/WebRTC.xcframework/ios-i386_x86_64-simulator/WebRTC.framework/Headers/WebRTC.h
@@ -32,6 +32,7 @@
 #import <WebRTC/RTCAudioSessionConfiguration.h>
 #import <WebRTC/RTCCameraVideoCapturer.h>
 #import <WebRTC/RTCFileVideoCapturer.h>
+#import <WebRTC/RTCNetworkMonitor.h>
 #import <WebRTC/RTCMTLVideoView.h>
 #import <WebRTC/RTCEAGLVideoView.h>
 #import <WebRTC/RTCVideoViewShading.h>
@@ -55,7 +56,6 @@
 #import <WebRTC/RTCFieldTrials.h>
 #import <WebRTC/RTCIceCandidate.h>
 #import <WebRTC/RTCIceServer.h>
-#import <WebRTC/RTCIntervalRange.h>
 #import <WebRTC/RTCLegacyStatsReport.h>
 #import <WebRTC/RTCMediaConstraints.h>
 #import <WebRTC/RTCMediaSource.h>
@@ -77,6 +77,7 @@
 #import <WebRTC/RTCDtmfSender.h>
 #import <WebRTC/RTCSSLAdapter.h>
 #import <WebRTC/RTCSessionDescription.h>
+#import <WebRTC/RTCStatisticsReport.h>
 #import <WebRTC/RTCTracing.h>
 #import <WebRTC/RTCCertificate.h>
 #import <WebRTC/RTCCryptoOptions.h>