+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#### Version 8.0.0
+* Update to WebRTC.framework M87
+* Update to WebRTC.framework M84
+
 #### Version 7.0.0
 * Update to WebRTC.framework M79
 * Update to WebRTC.framework M75